# OpenAPIDoku.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Film Management API",
    "description": "This API allows clients to interact with a film database, supporting operations such as fetching, creating, updating, and deleting film records."
  },
  "paths": {
    "/film/{id}": {
      "get": {
        "summary": "Find Film by ID",
        "description": "Returns a single film by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "ID of film to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Film"
                }
              }
            }
          },
          "404": {
            "description": "Film not found"
          }
        },
        "security": [
          {
            "filmUser": []
          }
        ]
      }
    },
    "/film": {
      "get": {
        "summary": "Get all films",
        "description": "Returns a list of all films in the database.",
        "responses": {
          "200": {
            "description": "A list of films",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Film"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "filmUser": []
          }
        ]
      },
      "post": {
        "summary": "Add a new film",
        "description": "Creates a new film in the database.",
        "requestBody": {
          "description": "Film to add",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Film"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Film created"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "filmAdmin": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "filmUser": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "/auth/login",
            "scopes": {
              "read": "Grants read access",
              "write": "Grants write access",
              "admin": "Grants access to admin operations"
            }
          }
        }
      },
      "filmAdmin": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "/auth/login",
            "scopes": {
              "admin": "Grants access to admin operations"
            }
          }
        }
      }
    },
    "schemas": {
      "Film": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titel": {
            "type": "string"
          },
          "regisseur": {
            "type": "string"
          },
          "erscheinungsdatum": {
            "type": "string",
            "format": "date"
          },
          "laufzeit": {
            "type": "integer"
          },
          "bewertung": {
            "type": "number",
            "format": "float"
          },
          "genre": {
            "$ref": "#/components/schemas/Genre"
          }
        }
      },
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bezeichnung": {
            "type": "string"
          }
        }
      }
    }
  }
}